pyproject.toml
[project]
name = "posting"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tenacity>=8.2.0",
]
